casted before

+++ sklearn/metrics/cluster/supervised.py
@@ -596,12 +596,12 @@ def mutual_info_score(labels_true, labels_pred, contingency=None):
                          type(contingency))
 
     contingency_sum = contingency.sum()
-    pi = np.ravel(contingency.sum(axis=1))
-    pj = np.ravel(contingency.sum(axis=0))
+    pi = np.ravel(contingency.sum(axis=1), dtype=np.int64)
+    pj = np.ravel(contingency.sum(axis=0), dtype=np.int64)
     log_contingency_nm = np.log(nz_val)
     contingency_nm = nz_val / contingency_sum
     # Don't need to calculate the full outer product, just for non-zeroes
-    outer = pi.take(nzx).astype(np.int64) * pj.take(nzy).astype(np.int64)
+    outer = pi.take(nzx) * pj.take(nzy)
     log_outer = -np.log(outer) + log(pi.sum()) + log(pj.sum())
     mi = (contingency_nm * (log_contingency_nm - log(contingency_sum)) +
           contingency_nm * log_outer)
